# backend/src/main/resources/openapi.yaml
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/accounts/{id}": {
      "put": {
        "tags": [
          "account-controller"
        ],
        "operationId": "updateAccount",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResponse"
                }
              }
            }
          }
        }
      }
    },
    "/accounts/default": {
      "get": {
        "tags": [
          "account-controller"
        ],
        "operationId": "getDefaultAccount",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "account-controller"
        ],
        "operationId": "updateDefaultAccount",
        "parameters": [
          {
            "name": "accountId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/rules": {
      "get": {
        "tags": [
          "rule-controller"
        ],
        "operationId": "getAllRules",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RuleResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "rule-controller"
        ],
        "operationId": "createRule",
        "parameters": [
          {
            "name": "applyToExistingTransactions",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RuleResponse"
                }
              }
            }
          }
        }
      }
    },
    "/mono/link": {
      "post": {
        "tags": [
          "mono-bank-controller"
        ],
        "operationId": "linkClient",
        "parameters": [
          {
            "name": "clientToken",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/mono/accounts/link": {
      "post": {
        "tags": [
          "mono-bank-controller"
        ],
        "operationId": "linkAccount",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkMonoBankAccountRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/login": {
      "get": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "login",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "loginWithCredentials",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/categories": {
      "get": {
        "tags": [
          "category-controller"
        ],
        "operationId": "getAllCategories",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "EXPENSE",
                "INCOME"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "category-controller"
        ],
        "operationId": "createCategory",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponse"
                }
              }
            }
          }
        }
      }
    },
    "/accounts": {
      "get": {
        "tags": [
          "account-controller"
        ],
        "operationId": "getAccounts",
        "parameters": [
          {
            "name": "skipBankAccounts",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AccountResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "account-controller"
        ],
        "operationId": "createAccount",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/password/encode": {
      "get": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "encodePassword",
        "parameters": [
          {
            "name": "password",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/categories/top-expenses": {
      "get": {
        "tags": [
          "category-controller"
        ],
        "operationId": "getTopExpenses",
        "parameters": [
          {
            "name": "accountId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "categoryIds",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "format": "uuid"
              }
            }
          },
          {
            "name": "categoryType",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "EXPENSE",
                "INCOME"
              ]
            }
          },
          {
            "name": "categoryIdsToExclude",
            "in": "query",
            "required": false,
            "schema": {
              "uniqueItems": true,
              "type": "array",
              "items": {
                "type": "string",
                "format": "uuid"
              }
            }
          },
          {
            "name": "dateFrom",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "dateTo",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TopExpenseResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/accounts/{accountId}": {
      "get": {
        "tags": [
          "account-controller"
        ],
        "operationId": "getAccount",
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AccountResponse": {
        "required": [
          "balance",
          "currency",
          "isBankAccount",
          "name",
          "type"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "balance": {
            "type": "number"
          },
          "currency": {
            "$ref": "#/components/schemas/CurrencyResponse"
          },
          "type": {
            "$ref": "#/components/schemas/AccountTypeResponse"
          },
          "isBankAccount": {
            "type": "boolean"
          }
        }
      },
      "AccountTypeResponse": {
        "required": [
          "fullName",
          "iconName",
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "fullName": {
            "type": "string"
          },
          "iconName": {
            "type": "string"
          }
        }
      },
      "CategoryResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "icon": {
            "type": "string"
          },
          "position": {
            "type": "integer",
            "format": "int32"
          },
          "type": {
            "type": "string",
            "enum": [
              "EXPENSE",
              "INCOME"
            ]
          },
          "subCategories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategoryResponse"
            }
          }
        }
      },
      "CurrencyResponse": {
        "required": [
          "code",
          "fullName",
          "shortName",
          "sign"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "fullName": {
            "type": "string"
          },
          "shortName": {
            "type": "string"
          },
          "sign": {
            "type": "string"
          }
        }
      },
...
      "AccountRequest": {
        "required": [
          "currencyCode",
          "name",
          "typeId"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "balance": {
            "type": "number"
          },
          "currencyCode": {
            "type": "integer",
            "format": "int32"
          },
          "typeId": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "RuleRequest": {
        "required": [
          "conditionText",
          "conditionType"
        ],
        "type": "object",
        "properties": {
          "conditionType": {
            "type": "string",
            "enum": [
              "TEXT_CONTAINS",
              "TEXT_EQUALS"
            ]
          },
          "conditionText": {
            "type": "string"
          },
          "assignCategoryId": {
            "type": "string",
            "format": "uuid"
          },
          "markAsTransferFromAccountId": {
            "type": "string",
            "format": "uuid"
          },
          "markAsTransferToAccountId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "AccountShortResponse": {
        "required": [
          "id",
          "name",
          "type"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/AccountTypeResponse"
          }
        }
      },
      "ConditionResponse": {
        "required": [
          "textToApply",
          "type"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "textToApply": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/ConditionTypeResponse"
          }
        }
      },
      "ConditionTypeResponse": {
        "required": [
          "description",
          "value"
        ],
        "type": "object",
        "properties": {
          "value": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "RuleResponse": {
        "required": [
          "condition",
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "condition": {
            "$ref": "#/components/schemas/ConditionResponse"
          },
          "assignCategory": {
            "$ref": "#/components/schemas/CategoryResponse"
          },
          "markAsTransferFromAccount": {
            "$ref": "#/components/schemas/AccountShortResponse"
          },
          "markAsTransferToAccount": {
            "$ref": "#/components/schemas/AccountShortResponse"
          }
        }
      },
      "LinkMonoBankAccountRequest": {
        "required": [
          "balance",
          "currencyCode",
          "iban",
          "id",
          "type"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "balance": {
            "type": "number"
          },
          "currencyCode": {
            "type": "integer",
            "format": "int32"
          },
          "maskedPan": {
            "type": "string"
          },
          "iban": {
            "type": "string"
          }
        }
      },
      "LoginRequest": {
        "required": [
          "password",
          "username"
        ],
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "AuthenticationResponse": {
        "required": [
          "token"
        ],
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          }
        }
      },
      "CategoryRequest": {
        "required": [
          "icon",
          "name",
          "position",
          "type"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "icon": {
            "type": "string"
          },
          "position": {
            "type": "integer",
            "format": "int32"
          },
          "type": {
            "type": "string",
            "enum": [
              "EXPENSE",
              "INCOME"
            ]
          }
        }
      },
...
      "TopExpenseResponse": {
        "required": [
          "total"
        ],
        "type": "object",
        "properties": {
          "total": {
            "type": "number"
          },
          "category": {
            "$ref": "#/components/schemas/CategoryResponse"
          }
        }
      },
...
    }
  }
}